# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Tarefas",
    "version": "1.0.0",
    "description": "API para gerenciamento de tarefas com autenticação JWT."
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "senha": { "type": "string" }
                },
                "required": ["email", "senha"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário registrado" },
          "400": { "description": "Dados inválidos" },
          "409": { "description": "Usuário já existe" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login do usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "senha": { "type": "string" }
                },
                "required": ["email", "senha"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login bem-sucedido" },
          "401": { "description": "Credenciais inválidas" }
        }
      }
    },
    "/tasks": {
      "post": {
        "summary": "Criar tarefa",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "status": { "type": "string" },
                  "assignee": { "type": "string" }
                },
                "required": ["title", "description"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Tarefa criada" },
          "400": { "description": "Dados inválidos" }
        }
      },
      "get": {
        "summary": "Listar tarefas",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Lista de tarefas" }
        }
      }
    },
    "/tasks/{id}": {
      "put": {
        "summary": "Atualizar tarefa",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "status": { "type": "string" },
                  "assignee": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Tarefa atualizada" },
          "404": { "description": "Tarefa não encontrada" }
        }
      },
      "delete": {
        "summary": "Deletar tarefa",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "204": { "description": "Tarefa deletada" },
          "404": { "description": "Tarefa não encontrada" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}